+++ packages/credential-sdk/src/types/accumulator/index.js
@@ -3,4 +3,4 @@ export * from './params';
 export * from './public-key';
 export * from './accumulator';
 export * from './counters';
-export * from './accumulator-id';
+export * from './id';