+++ packages/cheqd-blockchain-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @docknetwork/cheqd-blockchain-api
 
+## 0.28.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.36.0
+
 ## 0.28.0
 
 ### Minor Changes